requirements.txt
asgiref==3.2.3
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-import-export==2.0.1
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.3
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
